bench: move escape benchmarks to package-level
Moved encode/decode benchmarks from centralized benchmarks/ into each
package's bench/ directory. These are internal performance benchmarks,
not needed in the CodSpeed CI suite.
