+++ benchmarks/package.json
@@ -11,8 +11,6 @@
     "@codspeed/vitest-plugin": "^5.2.0",
     "@rethinkhealth/hl7v2": "workspace:*",
     "@rethinkhealth/hl7v2-builder": "workspace:*",
-    "@rethinkhealth/hl7v2-decode-escapes": "workspace:*",
-    "@rethinkhealth/hl7v2-encode-escapes": "workspace:*",
     "@rethinkhealth/hl7v2-mllp": "workspace:*",
     "@rethinkhealth/hl7v2-parser": "workspace:*",
     "@rethinkhealth/hl7v2-preset-lint-profile-recommended": "workspace:*",
